LayoutTests/http/tests/security/contentSecurityPolicy/block-mixed-content-hides-warning.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
This page should neither alert "FAIL" nor generate any mixed content warnings in the console.
<iframe src="https://127.0.0.1:8443/security/contentSecurityPolicy/resources/mixed-content-with-csp.html"></iframe>
